+++ src/Telemetry.h
@@ -9,12 +9,12 @@ private:
 public:
     Telemetry();
     String getState(byte State);
-    String constructMessage(float hour, float minute, float second,
+    String constructMessage(byte hour, byte minute, byte second,
                             float packetCount,
                             char mode, String getState, float altitudeBMP,
                             char HS_DEPLOYED, char PC_DEPLOYED, char MAST_RAISED,
                             float temperature, float batt, float pressure,
-                            float hour_ms, float minute_ms, float second_ms,
+                            byte hour_ms, byte minute_ms, byte second_ms,
                             float altitudeGPS, float latitude, float longitude,
                             byte satCount, float angleX, float angleY, String echo,
                             char buffer[256]
